Resize 1x1 images to 16x16
Minecraft is not smart about how it mipmaps textures. By default, it allows players to mipmap textures down by 4 powers of two, as the included textures are 16 by 16 (with the smallest mipmap being 2 by 2, or 2 to the power of 1). However, this behaviour can be disrupted in two ways. If a texture is introduced that is a lesser power of two (e.g. 8 by 8), Minecraft will lower the amount that it mipmaps to the amount that the smallest texture can mipmap to (e.g. for 8 by 8, Minecraft will only mipmap by 3 powers of 2). The other way this is disrupted is when a non-power of 2 sized texture is introduced (e.g. 5 by 7, 3 by 3, 1 by 1 etc). This will completely disable mipmaps for all textures.

The result of all of this is that if you use even only one texture with a size of a lower power of two than 2^4 (16 by 16) or a non power of two size (1 by 1 in this case), your game doesn’t use mipmaps.

TL;DR: Mojang wants all textures to be 16 by 16 or else they’ll take away (at least some of) your mipmaps.