Replace block of repeater texture with a cleaner edit
The official vanilla texture is 16 by 16, meaning that this is likely a cropped screenshot. I added the redstone torch head to the base texture of a repeater in the same position as vanilla, directly from the unlit redstone torch texture. I ran into a fair bit of frustration trying to generate this texture inside the code of the mod, so I'll probably make a pull request once I figure that out at another point. Doing that would mean the repeater texture would match up with a loaded recourse pack. The image's metadata has been stripped and the image has been compressed through the same process as the other images.